+++ ai_edge_torch/generative/quantize/example.py
@@ -26,7 +26,7 @@ def main():
   config = gemma.get_fake_model_config_2b_for_test()
   model = gemma.Gemma(config)
   idx = torch.from_numpy(np.array([[1, 2, 3, 4]]))
-  tokens = torch.full((1, 10), 0, dtype=torch.long, device="cpu")
+  tokens = torch.full((1, 10), 0, dtype=torch.int, device="cpu")
   tokens[0, :4] = idx
   input_pos = torch.arange(0, 10)
